fixed session persistence and security. Split functions and strava API into their own files.
